+++ CHANGELOG.md
@@ -8,3 +8,5 @@ CHANGELOG
 * Drop support of php5
 * Drop support of symfony/symfony < 3.4
 * Change bundle alias from `bo_shurik_telegram_bot` to `boshurik_telegram_bot`
+* Split `bin/console telegram:webhook` command to `bin/console telegram:webhook:set` 
+and `bin/console telegram:webhook:unset`